+++ tools-c++/src/FileMetadata.cc
@@ -1,3 +1,21 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include <iostream>
 #include <fstream>
 #include <vector>
@@ -130,7 +148,8 @@ int main(int argc, char* argv[])
 
   uint64_t paddedBytes = getTotalPaddingSize(footer);
   // empty ORC file is ~45 bytes. Assumption here is file length always >0
-  double percentPadding = static_cast<double>(paddedBytes) * 100 / fileSize;
+  double percentPadding =
+    static_cast<double>(paddedBytes) * 100 / static_cast<double>(fileSize);
   std::cout << "File length: " << fileSize << " bytes" << std::endl;
   std::cout <<"Padding length: " << paddedBytes << " bytes" << std::endl;
   std::cout <<"Padding ratio: " << std::fixed << std::setprecision(2)
